Refactor: modularize game logic
